Start `dev/lib` section
To get a list of cool libraries to use. Similar to stack section, but
less curated. Tech stack will be a section were I put stuff I really use
regularly. This is for a more comprehensive list.

Maybe there will be duplicates in the future: things that started on
`dev/lib` and were upgrade to `stack` once I started using them.